+++ packages/ui/src/lib/form/form.component.html.ts
@@ -6,27 +6,28 @@ export const formTemplate = `
     <div *ngIf="!wrapper.children.length&&!empty">
       <div data-grid>
         <div
-          [attr.data-col]="field?.getColumns()"
+          [attr.data-col]="getColumns(field)"
           *ngFor="let field of form?.fields | visibleFields:form"
-          [ngClass]="'ec-form-group_' + field?.view + ' ' + field?.classes"
-          [class.is-read-only]="formService.isReadOnly(field, form)"
           [ngSwitch]="field?.view"
-          class="form-group"
           [attr.data-type]="field?.type"
         >
-          <label
-            *ngIf="showLabel(field, form)"
-            [for]="field.id"
-            [title]="field.property"
-            >{{field.label || field.property}}</label
-          >
-          <ec-input
-            [field]="field"
-            [group]="group"
-            [item]="form"
-            [component]="field.getComponent('form')||field.input"
-            [formControl]="group.get(field.property)"
-          ></ec-input>
+          <div [ngClass]="'ec-field-group_' + field?.view + ' ' + field?.classes"
+          [class.is-read-only]="formService.isReadOnly(field, form)" class="field-group">
+            <label
+              *ngIf="showLabel(field, form)"
+              [for]="field.id"
+              [title]="field.property"
+              class="field-group__label"
+              >{{field.label || field.property}}</label
+            >
+            <ec-input
+              [field]="field"
+              [group]="group"
+              [item]="form"
+              [component]="field.getComponent('form')||field.input"
+              [formControl]="group.get(field.property)"
+            ></ec-input>
+          </div>
         </div>
       </div>
       <button
@@ -42,7 +43,7 @@ export const formTemplate = `
         >
       </button>
       <ec-loader
-        class="ec-loader ec-loader_overlay ec-loader_global"
+        class="ec-loader loader is-global"
         #formLoader
       ></ec-loader>
     </div>