+++ src/mocks/data.ts
@@ -212,8 +212,8 @@ export const mocked = {
             form: false,
             resolve: () => ' ',
             view: 'link',
-            class: 'btn btn_outlined',
-            icon: 'binoculars',
+            class: 'btn btn_minor btn_square',
+            icon: 'eye-open',
             action: (item, property) => {
               console.log('clicked button', item, property);
             },